requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
mcp==1.13.1
mdurl==0.1.2
openai==1.102.0
opentelemetry-api==1.36.0
pillow==11.3.0
protobuf==5.29.5
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.3
tiktoken==0.11.0
tqdm==4.67.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
